Reorganize procedure interaction level based test classes
