+++ enterprise/security/src/test/java/org/neo4j/server/security/enterprise/auth/EmbeddedBuiltInProceduresTest.java
@@ -21,11 +21,11 @@ package org.neo4j.server.security.enterprise.auth;
 
 import org.neo4j.kernel.enterprise.api.security.EnterpriseAuthSubject;
 
-public class NeoShallowEmbeddedScenariosTest extends AuthScenariosLogic<EnterpriseAuthSubject>
+public class EmbeddedBuiltInProceduresTest extends BuiltInProceduresInteractionTestBase<EnterpriseAuthSubject>
 {
     @Override
     protected NeoInteractionLevel<EnterpriseAuthSubject> setUpNeoServer() throws Throwable
     {
-        return new NeoShallowEmbeddedInteraction();
+        return new EmbeddedInteraction();
     }
 }
