+++ enterprise/security/src/test/java/org/neo4j/server/security/enterprise/auth/ProcedureInteractionTestBase.java
@@ -31,18 +31,22 @@ import java.util.function.Consumer;
 import java.util.stream.Stream;
 
 import org.neo4j.bolt.v1.transport.integration.Neo4jWithSocket;
+import org.neo4j.bolt.v1.transport.integration.TransportTestUtil;
+import org.neo4j.bolt.v1.transport.socket.client.SocketConnection;
+import org.neo4j.bolt.v1.transport.socket.client.TransportConnection;
 import org.neo4j.graphdb.GraphDatabaseService;
 import org.neo4j.graphdb.ResourceIterator;
-import org.neo4j.graphdb.Result;
+import org.neo4j.helpers.HostnamePort;
 import org.neo4j.kernel.api.security.exception.InvalidArgumentsException;
 import org.neo4j.kernel.impl.proc.Procedures;
 import org.neo4j.logging.Log;
 import org.neo4j.procedure.Context;
 import org.neo4j.procedure.Procedure;
 import org.neo4j.test.DoubleLatch;
+import org.neo4j.test.rule.concurrent.ThreadingRule;
 
+import static java.lang.String.format;
 import static java.util.stream.Collectors.toList;
-
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.either;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -51,33 +55,37 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.not;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.neo4j.bolt.v1.messaging.message.InitMessage.init;
+import static org.neo4j.bolt.v1.messaging.util.MessageMatchers.msgSuccess;
+import static org.neo4j.bolt.v1.transport.integration.TransportTestUtil.eventuallyReceives;
+import static org.neo4j.helpers.collection.MapUtil.map;
 import static org.neo4j.procedure.Procedure.Mode.READ;
 import static org.neo4j.procedure.Procedure.Mode.WRITE;
+import static org.neo4j.server.security.auth.AuthProceduresIT.assertKeyIsMap;
 import static org.neo4j.server.security.enterprise.auth.PredefinedRolesBuilder.ADMIN;
 import static org.neo4j.server.security.enterprise.auth.PredefinedRolesBuilder.ARCHITECT;
 import static org.neo4j.server.security.enterprise.auth.PredefinedRolesBuilder.PUBLISHER;
 import static org.neo4j.server.security.enterprise.auth.PredefinedRolesBuilder.READER;
 
-abstract class AuthTestBase<S>
+abstract class ProcedureInteractionTestBase<S>
 {
     protected boolean PWD_CHANGE_CHECK_FIRST = false;
     protected String CHANGE_PWD_ERR_MSG = AuthProcedures.PERMISSION_DENIED;
-    protected String BOLT_PWD_ERR_MSG =
+    private String BOLT_PWD_ERR_MSG =
             "The credentials you provided were valid, but must be changed before you can use this instance.";
-    protected String READ_OPS_NOT_ALLOWED = "Read operations are not allowed";
-    protected String WRITE_OPS_NOT_ALLOWED = "Write operations are not allowed";
-    protected String SCHEMA_OPS_NOT_ALLOWED = "Schema operations are not allowed";
+    String READ_OPS_NOT_ALLOWED = "Read operations are not allowed";
+    String WRITE_OPS_NOT_ALLOWED = "Write operations are not allowed";
+    String SCHEMA_OPS_NOT_ALLOWED = "Schema operations are not allowed";
 
     protected boolean IS_EMBEDDED = true;
-    protected boolean IS_BOLT = false;
+    boolean IS_BOLT = false;
 
-    protected String pwdReqErrMsg( String errMsg )
+    String pwdReqErrMsg( String errMsg )
     {
         return PWD_CHANGE_CHECK_FIRST ? CHANGE_PWD_ERR_MSG : IS_EMBEDDED ? errMsg : BOLT_PWD_ERR_MSG;
     }
 
-    final String EMPTY_ROLE = "empty";
+    private final String EMPTY_ROLE = "empty";
 
     S adminSubject;
     S schemaSubject;
@@ -90,7 +98,9 @@ abstract class AuthTestBase<S>
         "writeSubject", "pwdSubject", "noneSubject", "neo4j" };
     String[] initialRoles = { ADMIN, ARCHITECT, PUBLISHER, READER, EMPTY_ROLE };
 
-    protected EnterpriseUserManager userManager;
+    protected abstract ThreadingRule threading();
+
+    EnterpriseUserManager userManager;
 
     protected NeoInteractionLevel<S> neo;
 
@@ -138,7 +148,7 @@ abstract class AuthTestBase<S>
         return Stream.concat( Arrays.stream(strs), Arrays.stream( moreStr ) ).toArray( String[]::new );
     }
 
-    protected List<String> listOf( String... values )
+    List<String> listOf( String... values )
     {
         return Stream.of( values ).collect( toList() );
     }
@@ -321,7 +331,7 @@ abstract class AuthTestBase<S>
         assertThat( err, containsString( partOfErrorMsg ) );
     }
 
-    void assertFail( S subject, String call, String partOfErrorMsg1, String partOfErrorMsg2 )
+    private void assertFail( S subject, String call, String partOfErrorMsg1, String partOfErrorMsg2 )
     {
         String err = assertCallEmpty( subject, call );
         assertThat( err, not( equalTo( "" ) ) );
@@ -367,48 +377,50 @@ abstract class AuthTestBase<S>
         assertKeyIsArray( r, key, items );
     }
 
-    void assertKeyIsArray( ResourceIterator<Map<String, Object>> r, String key, String[] items )
+    private void assertKeyIsArray( ResourceIterator<Map<String,Object>> r, String key, String[] items )
     {
         List<Object> results = getObjectsAsList( r, key );
         assertEquals( Arrays.asList( items ).size(), results.size() );
         Assert.assertThat( results, containsInAnyOrder( items ) );
     }
 
-    protected void assertKeyIsMap( ResourceIterator<Map<String, Object>> r, String keyKey, String valueKey, Map<String,Object> expected )
+    // --------------------- helpers -----------------------
+
+    void shouldTerminateTransactionsForUser( S subject, String procedure ) throws Throwable
     {
-        List<Map<String, Object>> result = r.stream().collect( toList() );
+        DoubleLatch latch = new DoubleLatch( 2 );
+        ThreadedTransactionCreate<S> userThread = new ThreadedTransactionCreate<>( neo, latch );
+        userThread.execute( threading(), subject );
+        latch.startAndWaitForAllToStart();
 
-        assertEquals( "Results for should have size " + expected.size() + " but was " + result.size(),
-                expected.size(), result.size() );
+        assertEmpty( adminSubject, "CALL " + format(procedure, neo.nameOf( subject ) ) );
 
-        for ( Map<String, Object> row : result )
-        {
-            String key = (String) row.get( keyKey );
-            assertTrue( "Unexpected key '" + key + "'", expected.containsKey( key ) );
+        assertSuccess( adminSubject, "CALL dbms.security.listTransactions()",
+                r -> assertKeyIsMap( r, "username", "activeTransactions", map( "adminSubject", "1" ) ) );
 
-            assertTrue( "Value key '" + valueKey + "' not found in results", row.containsKey( valueKey ) );
-            Object objectValue = row.get( valueKey );
-            if ( objectValue instanceof List )
-            {
-                List<String> value = (List<String>) objectValue;
-                List<String> expectedValues = (List<String>) expected.get( key );
-                assertEquals( "Results for '" + key + "' should have size " + expectedValues.size() + " but was " +
-                        value.size(), value.size(), expectedValues.size() );
-                assertThat( value, containsInAnyOrder( expectedValues.toArray() ) );
-            }
-            else
-            {
-                String value = objectValue.toString();
-                String expectedValue = expected.get( key ).toString();
-                assertTrue(
-                        String.format( "Wrong value for '%s', expected '%s', got '%s'", key, expectedValue, value),
-                        value.equals( expectedValue )
-                    );
-            }
-        }
+        latch.finishAndWaitForAllToFinish();
+
+        userThread.closeAndAssertTransactionTermination();
+
+        assertEmpty( adminSubject, "MATCH (n:Test) RETURN n.name AS name" );
     }
 
-    public static class CountResult
+    @SuppressWarnings( "unchecked" )
+    TransportConnection startBoltSession( String username, String password ) throws Exception
+    {
+        TransportConnection connection = new SocketConnection();
+        HostnamePort address = new HostnamePort( "localhost:7687" );
+        Map<String,Object> authToken = map( "principal", username, "credentials", password, "scheme", "basic" );
+
+        connection.connect( address ).send( TransportTestUtil.acceptedVersions( 1, 0, 0, 0 ) )
+                .send( TransportTestUtil.chunk( init( "TestClient/1.1", authToken ) ) );
+
+        assertThat( connection, eventuallyReceives( new byte[]{0, 0, 0, 1} ) );
+        assertThat( connection, eventuallyReceives( msgSuccess() ) );
+        return connection;
+    }
+
+    private static class CountResult
     {
         public final String count;
 
@@ -418,7 +430,7 @@ abstract class AuthTestBase<S>
         }
     }
 
-    public static class ClassWithProcedures
+    static class ClassWithProcedures
     {
         @Context
         public GraphDatabaseService db;
@@ -490,7 +502,7 @@ abstract class AuthTestBase<S>
             Runnable runBefore;
             Runnable runAfter;
 
-            public LatchedRunnables( DoubleLatch doubleLatch, Runnable runBefore, Runnable runAfter )
+            LatchedRunnables( DoubleLatch doubleLatch, Runnable runBefore, Runnable runAfter )
             {
                 this.doubleLatch = doubleLatch;
                 this.runBefore = runBefore;
@@ -504,7 +516,7 @@ abstract class AuthTestBase<S>
             }
         }
 
-        public static void setTestLatch( LatchedRunnables testLatch )
+        static void setTestLatch( LatchedRunnables testLatch )
         {
             ClassWithProcedures.testLatch.set( testLatch );
         }
